pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "income-estimation"
version = "1.0.0"
description = "Income Estimation Models Analysis: Balancing Accuracy with Regulatory Requirements in the Georgian Banking Sector"
authors = [
    {name = "Sandro Gogaladze", email = "sandro.gogaladze@iset.ge"}
]
readme = "README.md"
license = {text = "Academic Use Only"}
requires-python = ">=3.9"

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "xgboost>=1.5.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "PyYAML>=6.0",
    "joblib>=1.0.0",
    "openpyxl>=3.0.0",
    "tqdm>=4.60.0",
    "shap>=0.40.0",
]

[project.optional-dependencies]
notebook = [
    "jupyter>=1.0.0",
    "ipywidgets>=7.6.0",
    "notebook>=6.4.0",
]
